add some basic component tests
jest and enzyme have been added to enable simple component and container
tests. In order to be able to test react components with async methods,
babel-preset-2015 was replaced with babel-preset-env.

The components are tested with snapshot-tests as well as more granular
shallow tests using enzyme. The containers are tested using enzyme shallow
tests.

The test can be invoked using "npm test".

This commit is a proposal for #10.